set background color white
